+++ sdk/dart/lib/src/cel_runtime_options.dart
@@ -1,16 +1,16 @@
-const packageVersion = '0.5.0';
+const packageVersion = '0.5.1';
 const wireProtocolVersion = 1;
 const _defaultWasmIntegrity =
-    'sha256-jrmsc98JwX7ZaSr7SfFYp3JKyEbLh6Oc2geq7aIqH0M=';
+    'sha256-m7oRta6ZTZiIK0knb+IC7TNTKKiSMzij5rU4QEbOpgo=';
 const _defaultWasmExecIntegrity =
     'sha256-DJSfSZb5qJaY5LXFht4yJJw7abe6rbZNIgBzzASsuhQ=';
 
 final class CelRuntimeOptions {
   const CelRuntimeOptions({
     this.wasmUrl =
-        'https://github.com/0xfe10/cel-bridge/releases/download/v0.5.0/cel_bridge.wasm',
+        'https://github.com/0xfe10/cel-bridge/releases/download/v0.5.1/cel_bridge.wasm',
     this.wasmExecUrl =
-        'https://github.com/0xfe10/cel-bridge/releases/download/v0.5.0/wasm_exec.js',
+        'https://github.com/0xfe10/cel-bridge/releases/download/v0.5.1/wasm_exec.js',
     this.wasmIntegrity = _defaultWasmIntegrity,
     this.wasmExecIntegrity = _defaultWasmExecIntegrity,
     this.profile,